Pulled out base ThreadAhead functionality into own class
since it's a useful functionality to add to other things.
